+++ packages/article-magazine-standard/src/article-lead-asset/article-lead-asset-prop-types.js
@@ -1,0 +1,11 @@
+import PropTypes from "prop-types";
+
+export const leadAssetPropTypes = {
+  caption: PropTypes.string,
+  credits: PropTypes.string
+};
+
+export const leadAssetDefaults = {
+  caption: null,
+  credits: null
+};