+++ android-app/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.10.37"></a>
+## [0.10.37](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.36...@times-components/android-app@0.10.37) (2018-12-03)
+
+**Note:** Version bump only for package @times-components/android-app
+
+
+
+
+
 <a name="0.10.36"></a>
 ## [0.10.36](https://github.com/newsuk/times-components/compare/@times-components/android-app@0.10.35...@times-components/android-app@0.10.36) (2018-12-03)